Allows CI on DuMuX master to fail
Changes on DuMuX master may be incompatible with previous releases. We
check against DuMuX master in the CI to be notified of such changes.
However, chacking against DuMuX master also leads failing nightly tests
often which give the impression that the released versions of the
adapter are broken even if this is not the case.

This change allows checks on DuMuX master to fail. A warning will be
created in these cases instead of an error.