Add CMake macrot for file links to preCICE files
A CMake macro called `add_precice_file_links` is added. If one
calls this macro in a `CMakeLists.txt` via `add_precice_file_links()`,
file links to all XML files in the directory of the `CMakeLists.txt`
are created.

The paths in tests have been adapted accordingly to use 
`CMAKE_CURRENT_BINARY_DIR` instead to refering to the source directory.