Move CI to GitHub actions (#2)
Moves the tests from GitLab CI to GitHub Actions

The following tests have been moved:
- Markdownlint
- black (Python linting)
- clang-format
- Build code

For more (time) efficent building and running the code I produced some Docker
containers that contain DUNE, DUMUX and preCICE. The building and updating
process of these containers should be improved and automatized.

The PR also updates the recipes for building the Docker containers. One can find
them in docker/ where also an additional README.md is located that describes
the recipes briefly. Once change is that the user inside the container is root to
simplify the testing process with GitHub.